Updates after Udemy course section 58 - async within DB trans
